Good Bro

+++ Services/User/IUserService.cs
@@ -1,4 +1,6 @@
 using SFM_BE.DTOs.Users;
+using System.Collections.Generic;
+using System.Threading.Tasks;
 
 namespace SFM_BE.Services.User;
 
@@ -6,9 +8,9 @@ public interface IUserService
 {
     Task<List<UserResponseDto>> GetUsersAsync();
 
-    Task<UserResponseDto> GetUserAsync(int id);
+    Task<UserResponseDto> GetUserAsync(long id);
 
-    Task UpdateAsync(int id, UpdateUserDto dto);
+    Task UpdateAsync(long id, UpdateUserDto dto);
 
-    Task DeleteAsync(int id);
+    Task DeleteAsync(long id);
 }
